+++ module/item/item.mjs
@@ -623,6 +623,11 @@ export class HeroSystem6eItem extends Item {
                 break;
         }
 
+        // DENSITYINCREASE can affect Encumbrance & Movements
+        if (this.system.XMLID === "DENSITYINCREASE") {
+            await this.actor.applyEncumbrancePenalty();
+        }
+
         // Charges expire
         // if (charges) {
         //     // Find the active effect
@@ -1412,6 +1417,25 @@ export class HeroSystem6eItem extends Item {
                     item.flags.tags.dcv = "";
                 }
                 item.flags.tags.dcv += `-5 Haymaker`;
+            }
+
+            // STRMINIMUM
+            const STRMINIMUM = item.findModsByXmlid("STRMINIMUM");
+            if (STRMINIMUM) {
+                const strMinimumValue = parseInt(STRMINIMUM.OPTION_ALIAS.match(/\d+/)?.[0] || 0);
+                const extraStr =
+                    Math.max(0, parseInt(item.actor?.system.characteristics.str.value || 0)) - strMinimumValue;
+                if (extraStr < 0) {
+                    const adjustment = Math.floor(extraStr / 5);
+                    item.system.ocvEstimated = `${parseInt(item.system.ocvEstimated) + adjustment}`;
+
+                    if (item.flags.tags.ocv) {
+                        item.flags.tags.ocv += "\n";
+                    } else {
+                        item.flags.tags.ocv = "";
+                    }
+                    item.flags.tags.ocv += `${adjustment.signedString()} ${STRMINIMUM.ALIAS}`;
+                }
             }
 
             item.system.phase = item.system.PHASE;
@@ -4868,6 +4892,11 @@ export class HeroSystem6eItem extends Item {
             return "PD";
         }
 
+        // STRIKE
+        if (this.system.EFFECT?.includes("STR")) {
+            return "PD";
+        }
+
         if (this.system.XMLID === "TELEKINESIS") {
             return "PD";
         }
@@ -4881,7 +4910,11 @@ export class HeroSystem6eItem extends Item {
             return "KB";
         }
 
-        console.error(`Unable to determine defense for ${this.name}`);
+        if (this.system.XMLID === "HANDTOHANDATTACK") {
+            return "PD";
+        }
+
+        console.warn(`Unable to determine defense for ${this.name}`);
         return "PD"; // Default
     }
 
@@ -4891,7 +4924,7 @@ export class HeroSystem6eItem extends Item {
         // A backpack from MiscEquipment.hdp is a CUSTOMPOWER
         if (this.system.description.match(/can hold \d+kg/i)) return true;
 
-        return this.baseInfo.isContainer;
+        return this.baseInfo?.isContainer;
     }
 
     get killing() {
@@ -4949,7 +4982,47 @@ export function getItem(id) {
     return null;
 }
 
-export async function RequiresASkillRollCheck(item, event) {
+export async function RequiresACharacteristicRollCheck(actor, characteristic, reasonText) {
+    console.log(characteristic, this);
+    const successValue = parseInt(actor?.system.characteristics[characteristic.toLowerCase()].roll) || 8;
+    const activationRoller = new HeroRoller().makeSuccessRoll(true, successValue).addDice(3);
+    await activationRoller.roll();
+    let succeeded = activationRoller.getSuccess();
+    const autoSuccess = activationRoller.getAutoSuccess();
+    const total = activationRoller.getSuccessTotal();
+    const margin = successValue - total;
+
+    const flavor = `${reasonText ? `${reasonText}. ` : ``}${characteristic.toUpperCase()} roll ${successValue}- ${
+        succeeded ? "succeeded" : "failed"
+    } by ${autoSuccess === undefined ? `${Math.abs(margin)}` : `rolling ${total}`}`;
+    let cardHtml = await activationRoller.render(flavor);
+
+    // FORCE success
+    if (!succeeded && overrideCanAct) {
+        const overrideKeyText = game.keybindings.get(HEROSYS.module, "OverrideCanAct")?.[0].key;
+        ui.notifications.info(`${actor.name} succeeded roll because override key.`);
+        succeeded = true;
+        cardHtml += `<p>Succeeded roll because ${game.user.name} used <b>${overrideKeyText}</b> key to override.</p>`;
+    }
+
+    const token = actor.token;
+    const speaker = ChatMessage.getSpeaker({ actor: actor, token });
+    speaker.alias = actor.name;
+
+    const chatData = {
+        type: CONST.CHAT_MESSAGE_TYPES.ROLL,
+        rolls: activationRoller.rawRolls(),
+        user: game.user._id,
+        content: cardHtml,
+        speaker: speaker,
+    };
+
+    await ChatMessage.create(chatData);
+
+    return succeeded;
+}
+
+export async function RequiresASkillRollCheck(item) {
     // Toggles don't need a roll to turn off
     //if (item.system?.active === true) return true;
 
@@ -5063,10 +5136,11 @@ export async function RequiresASkillRollCheck(item, event) {
         let cardHtml = await activationRoller.render(flavor);
 
         // FORCE success
-        if (!succeeded && event?.ctrlKey) {
-            ui.notifications.info(`${item.actor.name} succeeded roll because ${game.user.name} used CTRL key.`);
+        if (!succeeded && overrideCanAct) {
+            const overrideKeyText = game.keybindings.get(HEROSYS.module, "OverrideCanAct")?.[0].key;
+            ui.notifications.info(`${item.actor.name} succeeded roll because override key.`);
             succeeded = true;
-            cardHtml += `<p>Succeeded roll because ${game.user.name} used CTRL key.</p>`;
+            cardHtml += `<p>Succeeded roll because ${game.user.name} used <b>${overrideKeyText}</b> key to override.</p>`;
         }
 
         const actor = item.actor;